[backend] reorganize directory
