+++ pkg/router/router.go
@@ -13,6 +13,9 @@ import (
 	"github.com/jykuo-love-shiritori/twp/minio"
 	"github.com/jykuo-love-shiritori/twp/pkg/auth"
 	"github.com/jykuo-love-shiritori/twp/pkg/constants"
+	"github.com/jykuo-love-shiritori/twp/pkg/router/admin"
+	"github.com/jykuo-love-shiritori/twp/pkg/router/buyer"
+	"github.com/jykuo-love-shiritori/twp/pkg/router/general"
 )
 
 //	@title			twp API
@@ -59,16 +62,16 @@ func RegisterApi(e *echo.Echo, pg *db.DB, mc *minio.MC, logger *zap.SugaredLogge
 	api.POST("/oauth/logout", auth.Logout(pg, logger), auth.ValidateJwt(pg, logger))
 
 	// admin
-	api.GET("/admin/user", adminGetUser(pg, mc, logger))
-	api.DELETE("/admin/user/:username", adminDisableUser(pg, logger))
+	api.GET("/admin/user", admin.GetUser(pg, mc, logger))
+	api.DELETE("/admin/user/:username", admin.DisableUser(pg, logger))
 
-	api.GET("/admin/coupon", adminGetCoupon(pg, logger))
-	api.GET("/admin/coupon/:id", adminGetCouponDetail(pg, logger))
-	api.POST("/admin/coupon", adminAddCoupon(pg, logger))
-	api.PATCH("/admin/coupon/:id", adminEditCoupon(pg, logger))
-	api.DELETE("/admin/coupon/:id", adminDeleteCoupon(pg, logger))
+	api.GET("/admin/coupon", admin.GetCoupon(pg, logger))
+	api.GET("/admin/coupon/:id", admin.GetCouponDetail(pg, logger))
+	api.POST("/admin/coupon", admin.AddCoupon(pg, logger))
+	api.PATCH("/admin/coupon/:id", admin.EditCoupon(pg, logger))
+	api.DELETE("/admin/coupon/:id", admin.DeleteCoupon(pg, logger))
 
-	api.GET("/admin/report", adminGetReport(pg, mc, logger))
+	api.GET("/admin/report", admin.GetReport(pg, mc, logger))
 
 	// user
 	api.GET("/user/info", userGetInfo(pg, logger))
@@ -80,36 +83,36 @@ func RegisterApi(e *echo.Echo, pg *db.DB, mc *minio.MC, logger *zap.SugaredLogge
 	api.PATCH("/user/security/credit_card", userUpdateCreditCard(pg, logger))
 
 	// general
-	api.GET("/shop/:seller_name", getShopInfo(pg, mc, logger)) // user
-	api.GET("/shop/:seller_name/coupon", getShopCoupon(pg, logger))
-	api.GET("/shop/:seller_name/search", searchShopProduct(pg, mc, logger))
+	api.GET("/shop/:seller_name", general.GetShopInfo(pg, mc, logger)) // user
+	api.GET("/shop/:seller_name/coupon", general.GetShopCoupon(pg, logger))
+	api.GET("/shop/:seller_name/search", general.SearchShopProduct(pg, mc, logger))
 
-	api.GET("/tag/:id", getTagInfo(pg, logger))
+	api.GET("/tag/:id", general.GetTagInfo(pg, logger))
 
-	api.GET("/search", search(pg, mc, logger)) // search both product and shop
-	api.GET("/search/shop", searchShopByName(pg, mc, logger))
+	api.GET("/search", general.Search(pg, mc, logger)) // search both product and shop
+	api.GET("/search/shop", general.SearchShopByName(pg, mc, logger))
 
-	api.GET("/news", getNews(pg, logger))
-	api.GET("/news/:id", getNewsDetail(pg, logger))
-	api.GET("/discover", getDiscover(pg, mc, logger))
-	api.GET("/popular", getPopular(pg, mc, logger))
+	api.GET("/news", general.GetNews(pg, logger))
+	api.GET("/news/:id", general.GetNewsDetail(pg, logger))
+	api.GET("/discover", general.GetDiscover(pg, mc, logger))
+	api.GET("/popular", general.GetPopular(pg, mc, logger))
 
-	api.GET("/product/:id", getProductInfo(pg, mc, logger))
+	api.GET("/product/:id", general.GetProductInfo(pg, mc, logger))
 
 	// buyer
-	api.GET("/buyer/order", buyerGetOrderHistory(pg, mc, logger))
-	api.GET("/buyer/order/:id", buyerGetOrderDetail(pg, mc, logger))
+	api.GET("/buyer/order", buyer.GetOrderHistory(pg, mc, logger))
+	api.GET("/buyer/order/:id", buyer.GetOrderDetail(pg, mc, logger))
 
-	api.GET("/buyer/cart", buyerGetCart(pg, mc, logger)) // include product and coupon
-	api.GET("/buyer/coupon/:id", buyerGetCoupon(pg, logger))
-	api.POST("/buyer/cart/:cart_id/product/:product_id", buyerAddProductToCart(pg, logger))
-	api.POST("/buyer/cart/:cart_id/coupon/:coupon_id", buyerAddCouponToCart(pg, logger))
-	api.PATCH("buyer/cart/:cart_id/product/:product_id", buyerEditProductInCart(pg, logger))
-	api.DELETE("/buyer/:cart_id/product/:product_id", buyerDeleteProductFromCart(pg, logger))
-	api.DELETE("/buyer/cart/:cart_id/coupon/:coupon_id", buyerDeleteCouponFromCart(pg, logger))
+	api.GET("/buyer/cart", buyer.GetCart(pg, mc, logger)) // include product and coupon
+	api.GET("/buyer/cart/:cart_id/coupon", buyer.GetCoupon(pg, logger))
+	api.POST("/buyer/cart/:cart_id/product/:product_id", buyer.AddProductToCart(pg, logger))
+	api.POST("/buyer/cart/:cart_id/coupon/:coupon_id", buyer.AddCouponToCart(pg, logger))
+	api.PATCH("/buyer/cart/:cart_id/product/:product_id", buyer.EditProductInCart(pg, logger))
+	api.DELETE("/buyer/cart/:cart_id/product/:product_id", buyer.DeleteProductFromCart(pg, logger))
+	api.DELETE("/buyer/cart/:cart_id/coupon/:coupon_id", buyer.DeleteCouponFromCart(pg, logger))
 
-	api.GET("/buyer/cart/:cart_id/checkout", buyerGetCheckout(pg, logger))
-	api.POST("/buyer/cart/:cart_id/checkout", buyerCheckout(pg, logger))
+	api.GET("/buyer/cart/:cart_id/checkout", buyer.GetCheckout(pg, logger))
+	api.POST("/buyer/cart/:cart_id/checkout", buyer.Checkout(pg, logger))
 
 	// seller
 	api.GET("/seller/info", sellerGetShopInfo(pg, mc, logger))
